Optimized & Restructured SKV Mini App
