+++ lib/MiniApps/SkyView/Screens/DetailScreen/SKVDetailScreen.dart
@@ -1,0 +1,178 @@
+import 'dart:async';
+
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_uis/Widgets/Screen/Screen.dart';
+import 'package:flutter_uis/configs/AppDimensions.dart';
+
+import 'package:flutter_uis/Utils.dart';
+import 'package:flutter_uis/UI.dart';
+
+import 'widgets/SKVDetailScreenSpaceBackground.dart';
+import 'widgets/SKVDetailScreenTextContent.dart';
+import 'widgets/SKVDetailScreenPlanet.dart';
+import 'widgets/SKVDetailScreenOrbit.dart';
+import '../../configs/theme.dart' as theme;
+import '../../data/data.dart' as data;
+import 'Dimensions.dart';
+
+class SKVDetailScreen extends StatefulWidget {
+  final int index;
+  SKVDetailScreen(this.index, {Key key}) : super(key: key);
+
+  _SKVDetailScreenState createState() => _SKVDetailScreenState();
+}
+
+class _SKVDetailScreenState extends State<SKVDetailScreen>
+    with SingleTickerProviderStateMixin {
+  TabController tabController;
+  PageController pageController;
+  ScrollController starsController;
+
+  bool pageRendered = false;
+  int activePage;
+  Timer timeout;
+  GlobalKey starsWidgetKey = GlobalKey();
+
+  void initState() {
+    super.initState();
+    this.activePage = widget.index;
+    this.pageController = new PageController(
+      initialPage: widget.index,
+      keepPage: true,
+    );
+    this.starsController = new ScrollController();
+    this.pageController.addListener(() {
+      final offset = this.pageController.offset;
+      final screenWidth = AppDimensions.size.width;
+      final totalScroll = (data.objectList.length - 1) * screenWidth;
+      final starsWidget = this.starsWidgetKey.currentContext;
+
+      if (starsWidget != null) {
+        final starsOffset = starsWidget.size.width - screenWidth;
+        this.starsController.jumpTo(Utils.rangeMap(
+              offset,
+              0,
+              totalScroll,
+              0,
+              starsOffset,
+            ));
+        setState(() {});
+      }
+    });
+
+    Utils.lightStatusBar();
+
+    this.timeout = Timer(
+      Duration(milliseconds: 1300),
+      () => setState(() {
+        this.pageRendered = true;
+      }),
+    );
+  }
+
+  @override
+  void dispose() {
+    Utils.darkStatusBar();
+    this.pageController.dispose();
+    this.timeout.cancel();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    UI.init(context);
+    Dimensions.init(context);
+
+    final fontStyle = Theme.of(context).textTheme.body1.copyWith(
+          fontFamily: 'Montserrat',
+        );
+
+    final rootTheme = Theme.of(context).copyWith(
+      primaryColor: theme.secondary,
+      accentColor: theme.secondary,
+    );
+
+    return Screen(
+      Dimensions.init,
+      textStyle: fontStyle,
+      theme: rootTheme,
+      builder: (_) {
+        return RawKeyboardListener(
+          autofocus: true,
+          focusNode: FocusNode(),
+          onKey: (event) {
+            final key = event.logicalKey.debugName;
+            final rightKeys = ['Key K', 'Arrow Right'];
+            final leftKeys = ['Key I', 'Arrow Left'];
+            final duration =
+                (AppDimensions.size.width * 0.5).clamp(500.0, 800.0).toInt();
+
+            if (event.runtimeType.toString() == 'RawKeyUpEvent') {
+              if (rightKeys.contains(key) &&
+                  activePage < data.objectList.length - 1) {
+                this.pageController.animateToPage(
+                      this.activePage + 1,
+                      duration: Duration(milliseconds: duration),
+                      curve: Curves.linear,
+                    );
+              } else if (leftKeys.contains(key) && activePage > 0) {
+                this.pageController.animateToPage(
+                      this.activePage - 1,
+                      duration: Duration(milliseconds: duration),
+                      curve: Curves.linear,
+                    );
+              }
+            }
+          },
+          child: Stack(
+            fit: StackFit.expand,
+            children: <Widget>[
+              SKVDetailScreenSpaceBackground(
+                pageRendered: this.pageRendered,
+                starsWidgetKey: this.starsWidgetKey,
+                starsController: this.starsController,
+              ),
+              Positioned.fill(
+                child: PageView.builder(
+                  controller: this.pageController,
+                  itemCount: data.objectList.length,
+                  scrollDirection: Axis.horizontal,
+                  onPageChanged: (index) =>
+                      setState(() => (this.activePage = index)),
+                  pageSnapping: true,
+                  physics: this.pageRendered
+                      ? new ClampingScrollPhysics()
+                      : new NeverScrollableScrollPhysics(),
+                  itemBuilder: (ctx, index) {
+                    data.SKVObject item = data.objectList[index];
+                    return Stack(
+                      fit: StackFit.expand,
+                      children: [
+                        SKVDetailScreenOrbit(
+                          index: index,
+                          pageRendered: pageRendered,
+                          offset: this.pageController.offset,
+                        ),
+                        SKVDetailScreenPlanet(
+                          item: item,
+                          index: index,
+                          pageRendered: pageRendered,
+                          offset: this.pageController.offset,
+                        ),
+                        SKVDetailScreenTextContent(
+                          item: item,
+                          pageRendered: this.pageRendered,
+                        ),
+                      ],
+                    );
+                  },
+                ),
+              )
+            ],
+          ),
+        );
+      },
+    );
+  }
+}
