+++ lib/MiniApps/SkyView/data/storiesList.dart
@@ -1,38 +1,38 @@
-import 'SpaceStory.dart';
-export 'SpaceStory.dart';
+import '../models/SKVStory.dart';
+export '../models/SKVStory.dart';
 
 final storyList = [
-  new SpaceStory(
+  new SKVStory(
     "Andromeda",
     5.0,
     "The Andromeda Galaxy also known as Messier 31, M31, or NGC 224 and originally the Andromeda Nebula",
     "assets/ma-sk/andromeda.png",
   ),
-  new SpaceStory(
+  new SKVStory(
     "Hubble Space Telescope",
     4.0,
     "Breathtaking photos and science-changing discoveries from 25 years of exploration.",
     "assets/ma-sk/hubble-space-telescope.png",
   ),
-  new SpaceStory(
+  new SKVStory(
     "International Space Station",
     5.0,
     "The International Space Station (ISS) is a space station (habitable artificial satellite) in low Earth orbit.",
     "assets/ma-sk/international-space-station.png",
   ),
-  new SpaceStory(
+  new SKVStory(
     "Andromeda",
     5.0,
     "The Andromeda Galaxy also known as Messier 31, M31, or NGC 224 and originally the Andromeda Nebula",
     "assets/ma-sk/andromeda.png",
   ),
-  new SpaceStory(
+  new SKVStory(
     "Hubble Space Telescope",
     4.0,
     "Breathtaking photos and science-changing discoveries from 25 years of exploration.",
     "assets/ma-sk/hubble-space-telescope.png",
   ),
-  new SpaceStory(
+  new SKVStory(
     "International Space Station",
     5.0,
     "The International Space Station (ISS) is a space station (habitable artificial satellite) in low Earth orbit.",
