+++ lib/screens/UIDetail/data.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/services.dart';
 import 'package:flutter_vector_icons/flutter_vector_icons.dart';
 
 import 'package:flutter_uis/statics/models/UIItem.dart';
